+++ packages/hippobase/hippobase_storage_server/lib/hippobase_storage_server.dart
@@ -8,4 +8,6 @@ export 'src/providers/in_memory_storage_provider.dart';
 export 'src/providers/s3_storage_provider.dart';
 export 'src/storage_client.dart';
 export 'src/storage_download_stream.dart';
+export 'src/storage_native_download.dart';
+export 'src/storage_native_provider.dart';
 export 'src/storage_provider.dart';